structured the project
