+++ rsa/utils/__init__.py
@@ -1,0 +1,1 @@
+from .prime import get_prime
