+++ tests/test_parallel.py
@@ -1,10 +1,9 @@
 """Test for multiprocess prime generation."""
 
 import rsa.helpers as helpers_namespace
-import rsa.parallel
-import rsa.prime
+import rsa.utils as utils_namespace
 
 
 def test_parallel_primegen():
-    p = rsa.parallel.get_prime(1024, 3)
+    p = utils_namespace.get_prime(1024, 3)
     assert helpers_namespace.bit_size(p) == 1024
